+++ TropicsProject/Assets/Tropics/Scripts/ModCompat/!ModChecks.cs
@@ -9,6 +9,22 @@ using static R2API.DirectorAPI;
 
 namespace Tropics.ModChecks
 {
+    public class IsStarstorm2
+    {
+        private static bool? _enabled;
+
+        public static bool enabled
+        {
+            get
+            {
+                if (_enabled == null)
+                {
+                    _enabled = (BepInEx.Bootstrap.Chainloader.PluginInfos.ContainsKey("com.TeamMoonstorm.MSU") && BepInEx.Bootstrap.Chainloader.PluginInfos.ContainsKey("com.TeamMoonstorm"));
+                }
+                return (bool)_enabled;
+            }
+        }
+    }
     public class IsEnemiesReturns
     {
         private static bool? _enabled;