+++ TropicsProject/Assets/Tropics/CHANGELOG.md
@@ -1,2 +1,7 @@
-# 0.0.1
+# 1.0.1
+* Changed boss theme (Hydrophobia -> A Tempestuous Noise of Thunder and Lightning Heard)
+* Added a config option to enable EnemiesReturns' Archer Bugs on Sunset/Midnight Tropics (disabled by default)
+* Fixed manifest.json file linking to Wetland Downpour's GitHub page (oops)
+
+# 1.0.0
 * Initial Release